safari-534.53-branch/LayoutTests/http/tests/security/dataURL/xss-DENIED-from-data-url-sub-frame-2-level-expected.txt

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL http://127.0.0.1:8000/security/dataURL/xss-DENIED-from-data-url-sub-frame-2-level.html from frame with URL data:text/html,%3Chtml%3E%3Chead%3E%3Cscript%3Efunction%20test()%20{try%20{top.document.getElementById(%22accessMe%22).innerHTML%20=%20%22FAIL:%20Cross%20frame%20access%20from%20a%20data:%20URL%20inside%20another%20data:%20URL%20was%20allowed.%22;}%20catch%20(e)%20{}if%20(window.layoutTestController)layoutTestController.notifyDone();}%3C/script%3E%3C/head%3E%3Cbody%20onload=%22test();%22%3E%3Cp%3EInner-inner%20iframe.%3C/p%3E%3C/body%3E%3C/html%3E. Domains, protocols and ports must match.

This tests that a data: URL loaded in an iframe inside another data: URL loaded iframe doesn't have access to the main frame.


PASS: Cross frame access from a data: URL inside another data: URL was denied.

--------
Frame: 'aFrame'
--------
Inner iframe.



--------
Frame: '<!--framePath //aFrame/<!--frame0-->-->'
--------
Inner-inner iframe.
